--- client/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <title>starter-nodejs-rest</title>
-    </head>
-    <body>
-        Starter for a Node.js RESTful Application using MongoDB and JSON Web Token
-    </body>
-</html>
